# app/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "knot-api Documentation",
        "description": "knot-api",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8088",
            "description": "knot-api server"
        }
    ],
    "paths": {
        "/api/v1/auth/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "login",
                "operationId": "dec7c3a25e32283876bea9ba4e3885b0",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "description": "User email",
                                        "type": "string"
                                    },
                                    "password": {
                                        "description": "User password",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "login success"
                    },
                    "422": {
                        "description": "Unprocessable entity"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                }
            }
        },
        "/api/v1/auth/register": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "signup",
                "operationId": "dc21a9ca3cb420fd0cf58cc6541f9ad2",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "first_name",
                                    "last_name",
                                    "email",
                                    "password",
                                    "password_confirmation"
                                ],
                                "properties": {
                                    "firstName": {
                                        "description": "User first name",
                                        "type": "string"
                                    },
                                    "lastName": {
                                        "description": "User last name",
                                        "type": "string"
                                    },
                                    "email": {
                                        "description": "User email",
                                        "type": "string"
                                    },
                                    "password": {
                                        "description": "User password",
                                        "type": "string"
                                    },
                                    "password_confirmation": {
                                        "description": "User password confirmation",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Register Successfully"
                    },
                    "422": {
                        "description": "Unprocessable entity"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                }
            }
        },
        "/api/v1/card": {
            "post": {
                "tags": [
                    "Card"
                ],
                "summary": "store new card",
                "operationId": "10aa39e9a858fafb9d0dc09fa5625d19",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "card_number",
                                    "cvv",
                                    "expiration"
                                ],
                                "properties": {
                                    "card_number": {
                                        "description": "Card Number",
                                        "type": "string"
                                    },
                                    "cvv": {
                                        "description": "Card CVV",
                                        "type": "string"
                                    },
                                    "expiration": {
                                        "description": "card expiry",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Card added Successfully"
                    },
                    "422": {
                        "description": "Unprocessable entity"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/merchant": {
            "get": {
                "tags": [
                    "Merchant"
                ],
                "summary": "List all merchants",
                "operationId": "a608f7aa03aeff395d09b6fd57a75ae9",
                "responses": {
                    "200": {
                        "description": "merchants retrieved success"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "not found"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Merchant"
                ],
                "summary": "add a new merchant",
                "operationId": "bfb9a2d9bc112652fbc1d4f5ef31b0ca",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "website"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "Merchant name",
                                        "type": "string"
                                    },
                                    "website": {
                                        "description": "Merchant Website",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Merchant added Successfully"
                    },
                    "422": {
                        "description": "Unprocessable entity"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "scheme": "bearer"
            }
        }
    }
}